requirements.txt
fpdf2
colorutils